chore: move demo to BasicDemo
