+++ demos/components/AddItemButton.vue
@@ -1,0 +1,5 @@
+<template>
+  <button class="tw-rounded tw-bg-blue-600 tw-mb-2 tw-text-xs tw-text-white tw-py-1 tw-px-2">Add random item</button>
+</template>
+
+<script lang="ts" setup></script>
